+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog (niondata)
 ====================
 
+15.6.2 (UNRELEASED)
+-------------------
+- Add power function.
+- Add radial profile function.
+- Improve handing of invalid inverse coordinates (1/0).
+
 15.6.1 (2023-10-23)
 -------------------
 - Minor update for typing compatibility.